requirements.txt
# Requirements
eel
PyCryptoDome
win10toast # not essential